backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.13",
    "greenlet>=3.2.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
